refactor to reduce duplicate code and more pylint fixes
